libpcl: optimize co_exit()
It was using co_exit_to() that creates the helper co-routine
that then deletes the exiting one. That helper co-routine was
running on a static memory area therefore not requiring the delete.
This whole sequence is expensive, as the co-routine creation
and switching all does the sigprocmask() syscall.

With this change the caller deletes the exited co-routine.
co_exit_to() with such technique is unsafe, so removed.