configure: add --disable-system-wa option
This option is trying to disable the system-specific work-arounds
(kernel-specific for now) that are needed for some particular (old)
versions of system software, different (older) than the ones found
on a build system.
The work-arounds that are needed on a build system itself, will
not be disabled but will print a compile-time warning.

This option is only interested for developers.
Users should not use it because the work-arounds can probe itself
and disable at run-time.
Enabling this option may prevent the same binary from working on
other systems.