signal: initially block all async signals also for sas wa
Blocking them only for sigreturn wa is not enough: if sigreturn
works properly but sigaltstack not, we still need to have them
initialli blocked to avoid the race when sas wa changed the stack
but have not yet disabled sas.

So we can have fatal signals initially unblocked only if both
work-arounds are unneeded.